some changes to clean stuff
