+++ templates/peers.html
@@ -1,0 +1,188 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>Mod Sharing Network</title>
+    <style>
+        body {
+            font-family: Arial, sans-serif;
+            max-width: 800px;
+            margin: 0 auto;
+            padding: 20px;
+            background-color: #f5f5f5;
+        }
+        h1 {
+            color: #333;
+            text-align: center;
+        }
+        .upload-section, .peer-section, .file-list {
+            background: white;
+            padding: 20px;
+            border-radius: 5px;
+            margin-bottom: 20px;
+            box-shadow: 0 2px 4px rgba(0,0,0,0.1);
+        }
+        table {
+            width: 100%;
+            border-collapse: collapse;
+        }
+        th, td {
+            padding: 8px;
+            text-align: left;
+            border-bottom: 1px solid #ddd;
+        }
+        tr:hover {
+            background-color: #f5f5f5;
+        }
+        button {
+            background-color: #4CAF50;
+            color: white;
+            padding: 6px 12px;
+            border: none;
+            border-radius: 4px;
+            cursor: pointer;
+        }
+        button:hover {
+            background-color: #45a049;
+        }
+        .file-size {
+            white-space: nowrap;
+        }
+        
+        /* Peer list styles */
+        .peer-category {
+            margin-top: 15px;
+            border: 1px solid #ddd;
+            border-radius: 5px;
+            overflow: hidden;
+        }
+        .peer-header {
+            padding: 10px;
+            background-color: #f0f0f0;
+            cursor: pointer;
+            display: flex;
+            justify-content: space-between;
+            align-items: center;
+        }
+        .peer-header h3 {
+            margin: 0;
+        }
+        .peer-list {
+            max-height: 0;
+            overflow: hidden;
+            transition: max-height 0.3s ease-out;
+        }
+        .peer-list.expanded {
+            max-height: 500px;
+            transition: max-height 0.5s ease-in;
+        }
+        .peer-item {
+            padding: 8px 15px;
+            border-bottom: 1px solid #eee;
+            display: flex;
+            align-items: center;
+        }
+        .peer-item:last-child {
+            border-bottom: none;
+        }
+        .friend {
+            color: #2e7d32;
+            background-color: #e8f5e9;
+        }
+        .peer {
+            color: #1565c0;
+            background-color: #e3f2fd;
+        }
+        .stranger {
+            background-color: #f5f5f5;
+        }
+        .enemy {
+            color: #c62828;
+            background-color: #ffebee;
+        }
+        .peer-id {
+            font-family: monospace;
+            margin-right: 10px;
+        }
+        .toggle-icon {
+            font-weight: bold;
+        }
+    </style>
+</head>
+<body>
+    <h1>Decentralized Mod Sharing</h1>
+    
+    <div class="peer-section">
+        <h2>Network Peers</h2>
+        <form method="post" action="/add_peer">
+            <input type="text" name="peer_address" placeholder="http://peer-address:port" required>
+            <button type="submit">Add Peer</button>
+        </form>
+        
+        {% if peer_types %}
+            {% for type, peers in peer_types.items() if peers %}
+            <div class="peer-category">
+                <div class="peer-header" onclick="togglePeerList('{{ type }}')">
+                    <h3>{{ type|capitalize }} ({{ peers|length }})</h3>
+                    <span class="toggle-icon" id="toggle-{{ type }}">+</span>
+                </div>
+                <div class="peer-list" id="list-{{ type }}">
+                    {% for peer in peers %}
+                    <div class="peer-item {{ type[:-1] }}">
+                        <span class="peer-id">{{ peer.split('//')[-1] }}</span>
+                        {% if type == 'friends' %}
+                        <form method="post" action="/set_peer_type" style="margin-left: auto;">
+                            <input type="hidden" name="peer_address" value="{{ peer }}">
+                            <input type="hidden" name="peer_type" value="peer">
+                            <button type="submit" style="background-color: #2196F3;">Demote to Peer</button>
+                        </form>
+                        {% elif type == 'peers' %}
+                        <form method="post" action="/set_peer_type" style="margin-left: auto;">
+                            <input type="hidden" name="peer_address" value="{{ peer }}">
+                            <input type="hidden" name="peer_type" value="friend">
+                            <button type="submit" style="background-color: #4CAF50;">Promote to Friend</button>
+                        </form>
+                        {% elif type == 'strangers' %}
+                        <form method="post" action="/set_peer_type" style="margin-left: auto;">
+                            <input type="hidden" name="peer_address" value="{{ peer }}">
+                            <input type="hidden" name="peer_type" value="peer">
+                            <button type="submit" style="background-color: #2196F3;">Promote to Peer</button>
+                        </form>
+                        {% endif %}
+                        {% if type != 'enemies' %}
+                        <form method="post" action="/set_peer_type" style="margin-left: 5px;">
+                            <input type="hidden" name="peer_address" value="{{ peer }}">
+                            <input type="hidden" name="peer_type" value="enemy">
+                            <button type="submit" style="background-color: #f44336;">Block</button>
+                        </form>
+                        {% else %}
+                        <form method="post" action="/set_peer_type" style="margin-left: auto;">
+                            <input type="hidden" name="peer_address" value="{{ peer }}">
+                            <input type="hidden" name="peer_type" value="stranger">
+                            <button type="submit" style="background-color: #9e9e9e;">Unblock</button>
+                        </form>
+                        {% endif %}
+                    </div>
+                    {% endfor %}
+                </div>
+            </div>
+            {% endfor %}
+        {% else %}
+            <p>No peers connected yet. Add a peer to join the network.</p>
+        {% endif %}
+    </div>
+    <script>
+        function togglePeerList(type) {
+            const list = document.getElementById(`list-${type}`);
+            const toggle = document.getElementById(`toggle-${type}`);
+            
+            if (list.classList.contains('expanded')) {
+                list.classList.remove('expanded');
+                toggle.textContent = '+';
+            } else {
+                list.classList.add('expanded');
+                toggle.textContent = '-';
+            }
+        }
+    </script>
+</body>
+</html>
